# openapi.yaml
openapi: 3.1.0
info:
  title: Bakame API
  version: 1.0.0
  description: |
    Bakame - Production-grade AI platform API

    All endpoints except /health require Bearer token authentication.

    Response Format:
    - Success: { data: <payload>, meta: { requestId: string } }
    - Error: { error: { code: string, message: string, requestId: string } }
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /chats:
    get:
      tags: [Chats]
      summary: List chats
      operationId: listChats
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: cursor
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
      responses:
        '200':
          description: Paginated list of chats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedChatList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Chats]
      summary: Create chat
      operationId: createChat
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Chat created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chats/{id}:
    get:
      tags: [Chats]
      summary: Get chat
      operationId: getChat
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChatId'
      responses:
        '200':
          description: Chat details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Chats]
      summary: Update chat
      operationId: updateChat
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChatId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        '200':
          description: Chat updated
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Chats]
      summary: Archive chat
      operationId: archiveChat
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChatId'
      responses:
        '204':
          description: Chat archived
        '404':
          $ref: '#/components/responses/NotFound'
  /chats/{id}/messages:
    get:
      tags: [Chats]
      summary: Get chat messages
      operationId: getMessages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChatId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated messages
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Chats]
      summary: Add message
      operationId: addMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChatId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 32000
      responses:
        '201':
          description: Message added
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ChatId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - PERMISSION_DENIED
                - NOT_FOUND
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
            requestId:
              type: string
    ChatResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            userId:
              type: string
            title:
              type: string
              nullable: true
            status:
              type: string
              enum: [active, archived]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/Meta'
    PaginatedChatList:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                    nullable: true
                  status:
                    type: string
                  messageCount:
                    type: integer
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
            nextCursor:
              type: string
              nullable: true
            hasMore:
              type: boolean
        meta:
          $ref: '#/components/schemas/Meta'
...
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string